Worked on tests.

+++ tests/test_ewfexport.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 # Export tool testing script
 #
-# Version: 20160318
+# Version: 20160320
 
 EXIT_SUCCESS=0;
 EXIT_FAILURE=1;
@@ -11,7 +11,7 @@ TEST_PREFIX=`pwd`;
 TEST_PREFIX=`dirname ${TEST_PREFIX}`;
 TEST_PREFIX=`basename ${TEST_PREFIX} | sed 's/^lib//'`;
 
-OPTION_SETS="raw ewf encase1 encase2 encase3 encase4 encase5 encase6 encase7 linen5 linen6 linen7 ftk smart ewfx encase7-v2";
+OPTION_SETS="format:encase1 format:encase2 format:encase3 format:encase4 format:encase5 format:encase6 format:encase7 format:encase7-v2 format:ewf format:ewfx format:ftk format:linen5 format:linen6 format:linen7 format:raw format:smart deflate:none deflate:empty-block deflate:fast deflate:best blocksize:16 blocksize:32 blocksize:128 blocksize:256 blocksize:512 blocksize:1024 blocksize:2048 blocksize:4096 blocksize:8192 blocksize:16384 blocksize:32768 hash:sha1 hash:sha256 hash:all";
 INPUT_GLOB="*.[Ees]*01";
 
 list_contains()
@@ -56,10 +56,11 @@ run_test()
 
 	if test -z "${OPTION_SET}";
 	then
-		OPTIONS="";
+		OPTIONS=();
 		TEST_OUTPUT="${INPUT_NAME}";
 	else
-		OPTIONS=`cat "${TEST_SET_DIR}/${INPUT_NAME}.${OPTION_SET}" | head -n 1 | sed 's/[\r\n]*$//'`;
+		OPTIONS_STRING=`cat "${TEST_SET_DIR}/${INPUT_NAME}.${OPTION_SET}" | head -n 1 | sed 's/[\r\n]*$//'`;
+		IFS=" " read -a OPTIONS <<< ${OPTIONS_STRING};
 		TEST_OUTPUT="${INPUT_NAME}-${OPTION_SET}";
 	fi
 	TMPDIR="tmp$$";
@@ -72,22 +73,27 @@ run_test()
 
 	# Note that options should not contain spaces otherwise the test_runner
 	# will fail parsing the arguments.
-	${TEST_RUNNER} ${TMPDIR} ${TEST_EXECUTABLE} -q -t ${TMPDIR}/${INPUT_NAME}.export -u ${OPTIONS} ${INPUT_FILE} > /dev/null;
+	${TEST_RUNNER} ${TMPDIR} ${TEST_EXECUTABLE} -q -t ${TMPDIR}/${INPUT_NAME}.export -u ${OPTIONS[*]} ${INPUT_FILE} > /dev/null;
 
 	RESULT=$?;
 
-	(cd ${TMPDIR} && md5sum ${INPUT_NAME}.export.* | sort -k 2 > ${TEST_OUTPUT}.log);
-
-	if test -f "${STORED_TEST_RESULTS}";
+	if test -f ${TMPDIR}/${INPUT_NAME}.export.raw;
 	then
-		zdiff ${STORED_TEST_RESULTS} ${TEST_RESULTS};
+		(cd ${TMPDIR} && md5sum ${INPUT_NAME}.export.* | sort -k 2 > ${TEST_OUTPUT}.log);
 
-		RESULT=$?;
-	else
-		gzip ${TEST_RESULTS};
+		if test -f "${STORED_TEST_RESULTS}";
+		then
+			zdiff ${STORED_TEST_RESULTS} ${TEST_RESULTS};
 
-		mv "${TEST_RESULTS}.gz" ${TEST_SET_DIR};
+			RESULT=$?;
+		else
+			gzip ${TEST_RESULTS};
+
+			mv "${TEST_RESULTS}.gz" ${TEST_SET_DIR};
+		fi
 	fi
+	# TODO: use ewfverify to verify non-raw exported files.
+
 	rm -rf ${TMPDIR};
 
 	if test -z "${OPTION_SET}";
